+++ src/core/plugins/plugins-selectors.ts
@@ -68,16 +68,11 @@ export async function waitForPlugins(ai: ApiInput): Promise<void> {
     const { init, locked } = props.state.plugins
     if (!locked) return
 
-    const { currency, rate, swap } = props.state.plugins
+    const { currency, swap } = props.state.plugins
     const missingPlugins: string[] = []
     for (const pluginId in init) {
       const shouldLoad = init[pluginId] !== false && init[pluginId] != null
-      if (
-        shouldLoad &&
-        currency[pluginId] == null &&
-        rate[pluginId] == null &&
-        swap[pluginId] == null
-      ) {
+      if (shouldLoad && currency[pluginId] == null && swap[pluginId] == null) {
         missingPlugins.push(pluginId)
       }
     }